Initial restructuring of code
