+++ src/main.rs
@@ -1,5 +1,9 @@
-mod app;
+mod components;
+mod store;
+mod types;
+mod utils;
+mod views;
 
 fn main() {
-    yew::Renderer::<app::App>::new().render();
+    yew::Renderer::<views::app::App>::new().render();
 }
